Fixed order of NiTriShapes in werewolf head mesh
Bloodmoon's werewolf head mesh has shapekeys like normal heads, but they are not on the first NiTriShape so text keys are not used for them. This leads to strange behavior when playing as a werewolf, such as the shapekeys looping endlessly despite their loop mode being set to clamp or not being played at all depending on the player's stance. The new head.nif fixes this, so a transformed player blinks normally and doesn't constantly open and close their mouth. To my knowledge, this does mean that the Talk animation group never plays, but it should not have been playing anyways.