add LODs to load doors
Adds LODs to the following meshes. In most cases, LOD 1 decreases the number of materials to one, using vertex color tints and UV manipulations that mimic the appearance of the multiple materials in LOD 0. For doors with a lot of triangles, LOD 2 simplifies them slightly to retain the overall visual appearance. Otherwise, LOD 2 or 3 reduces the door to the bare minimum number of triangles that remains door-shaped at extremely long view distances.

MGE XE distant land uses whichever LOD is visible at a distance of exactly 8192. I chose whether to put the LOD 1/2 boundary before or after the MGE XE cutoff based on how things looked in Nifskope at morrowind's default FOV (60). I expect that we will need to make adjustments to the LOD extents based on playtester feedback.

- bm_ic_door_01.nif
- bm_ic_door_pelt.nif
- bm_ic_door_pelt_dark.nif
- ex_cave_door_01.nif
- ex_colony_bardoor.nif
- ex_common_door_01.nif
- ex_common_door_balcony.nif
- ex_dae_door_load_oval.nif
- ex_de_ship_door.nif
- ex_de_ship_trapdoor.nif
- ex_emp_tower_01_a.nif
- ex_emp_tower_01_b.nif
- ex_h_trapdoor_01.nif
- ex_imp_loaddoor_01.nif
- ex_imp_loaddoor_02.nif
- ex_imp_loaddoor_03.nif
- ex_mh_sewer_trapdoor_01.nif
- ex_nord_door_01.nif
- ex_nord_door_02.nif
- ex_r_trapdoor_01.nif
- ex_redoran_barracks_01_a.nif
- ex_redoran_hut_01_a.nif
- ex_s_door.nif
- ex_s_door_double.nif
- ex_s_door_rounded.nif
- ex_t_door_01.nif
- ex_t_door_02.nif
- ex_t_door_stone_large.nif
- ex_v_cantondoor_01.nif
- ex_v_palace_grate_01.nif
- ex_v_palace_grate_02.nif
- ex_velothi_entrance_01_a.nif
- ex_velothi_entrance_03_a.nif
- ex_velothi_loaddoor_01.nif
- ex_velothi_loaddoor_02.nif
- hlaalu_loaddoor_ 01.nif
- hlaalu_loaddoor_ 02.nif